+++ R/dataframe-group_by-general.R
@@ -1,11 +1,12 @@
 # The env for storing group_by methods
 polars_groupby__methods <- new.env(parent = emptyenv())
 
-wrap_to_group_by <- function(x, by, maintain_order) {
+wrap_to_group_by <- function(x, by, maintain_order, predicates) {
   self <- new.env(parent = emptyenv())
   self$df <- x
   self$by <- by
   self$maintain_order <- maintain_order
+  self$predicates <- predicates
 
   class(self) <- c("polars_group_by", "polars_object")
   self
@@ -35,11 +36,18 @@ wrap_to_group_by <- function(x, by, maintain_order) {
 #'   c_mean_squared = (pl$col("c") ** 2)$mean()
 #' )
 groupby__agg <- function(...) {
-  self$df$lazy()$group_by(
-    !!!self$by,
-    .maintain_order = self$maintain_order
-  )$agg(...)$collect(optimizations = QueryOptFlags()$no_optimizations()) |>
-    wrap()
+  wrap({
+    out <- self$df$lazy()$group_by(
+      !!!self$by,
+      .maintain_order = self$maintain_order
+    )
+
+    if (!is.null(self$predicates)) {
+      out <- out$having(!!!self$predicates)
+    }
+
+    out$agg(...)$collect(optimizations = QueryOptFlags()$no_optimizations())
+  })
 }
 
 #' @inherit lazygroupby__head title params
@@ -226,4 +234,22 @@ groupby__len <- function(name = NULL) {
     }
     self$agg(len_expr)
   })
+}
+
+#' @inherit lazygroupby__having title description params
+#' @inherit dataframe__group_by return
+#' @examples
+#' df <- pl$DataFrame(x = c("a", "b", "a", "b", "c"))
+#'
+#' # Only keep groups that contain more than one element:
+#' df$group_by("x")$having(
+#'   pl$len() > 1
+#' )$agg()
+groupby__having <- function(...) {
+  wrap_to_group_by(
+    self$df,
+    by = self$by,
+    maintain_order = self$maintain_order,
+    predicates = c(self$predicates, list2(...))
+  )
 }